+++ docs/about/history.rst
@@ -36,6 +36,12 @@ Changed
       Useful if using *Water Body > Override Material* and still want an ocean.
    -  Improve multiple *Water Body* overlapping case when *Water Body > Override Material* option is used.
    -  Water Body adds an inclusion to clipping (ie unclips) if *Default Clipping State* is *Everything Clipped*.
+   -  Add scroll bar to *Ocean Debug GUI* when using *Draw LOD Datas Actual Size*.
+   -  Add support for *TrailRenderer*, *LineRenderer* and *ParticleSystem* to be used as ocean inputs in addition to *MeshRenderer*.
+   -  Un-deprecate *ShapeGerstner* as it is useful in some situations for adding a small number of distinct waves with high degree of control.
+   -  Add *Reverse Wave Weight* setting to *ShapeGerstner* for fine control over generated wave pairs.
+   -  Double sample count for *ShapeGerstner* waves to improve quality.
+   -  Add option (enabled by default) to prewarm foam simulation on load and camera teleports.
 
 Fixed
 ^^^^^